+++ changelog.md
@@ -1,5 +1,13 @@
 # Updates changelog
 
+SUN, 05-14-2017
+1. Fresh rebuild of entire project.
+2. New camera interface.
+3. Added camera flash light (if supported).
+4. Added camera swap from/to front/back (if supported).
+5. Added camera support for autofocus, auto white balance and auto exposure (if supported).
+6. Minor graphical & performance improvements.
+___
 SAT, 04-01-2017
 1. Fixed RecyclerView scrolling issue.
 2. Addedd Landscape Activity support.